Code update to be more engine friendly etc.
- Update InBetween()
- Moved MathExtensions.ToIntegral(..) to new Maths utility class
- Fixed some redudancies in StringExtensions.Parse()
- Fixed minor bug in StringExtensions.GetLocations()
- GenerateStream constant use instead of ""
- Added ability to parse fen from Span<string>
- Zobrist initialize cleaned up
- Added new Clock struct to contain time and inc from UCI
- SearchParameters and dependencies now uses ulong for integral values as default
- Expanded SearchParameters to include Depth/Nodes/Mate/Ponder flag
- Slight improvement to UCI interface
- Added overload to generate bitboard from one/two/... squares
- Added basic new Depth struct
- Improved Board.Squares() slightly
- Added UCI and SearchParameter properties to Game class
- Position now contains Searcher index (for multi-threaded searching etc)
- Added StringExtensions tests
- Fixed up PiecesTests

+++ Chess/Protocol/UCI/MovesToGoModel.cs
@@ -40,7 +40,7 @@ public sealed class MovesToGoModel : EventArgs, IMovesToGoModel
         MovesToGo = original.MovesToGo;
     }
 
-    public int MovesToGo { get; set; }
+    public ulong MovesToGo { get; set; }
 
     public ulong WhiteTimeMilliseconds
     {
